Conflicts:
	sa/profiles/DLink/DxS/get_version.py
	services/datasource/datasources/base.py

+++ lib/app/extmodelapplication.py
@@ -2,11 +2,12 @@
 # ----------------------------------------------------------------------
 # ExtModelApplication implementation
 # ----------------------------------------------------------------------
-# Copyright (C) 2007-2017 The NOC Project
+# Copyright (C) 2007-2018 The NOC Project
 # See LICENSE for details
 # ----------------------------------------------------------------------
 
 # Python modules
+from __future__ import absolute_import
 import datetime
 from functools import reduce
 # Third-party modules
@@ -19,17 +20,18 @@ from django.db.utils import IntegrityError
 from django.core.exceptions import ValidationError
 import six
 # NOC modules
-from extapplication import ExtApplication, view
 from noc.sa.interfaces.base import (
     BooleanParameter, IntParameter,
     FloatParameter, TagsParameter,
     NoneParameter, StringListParameter,
     DictParameter, ListOfParameter,
     ModelParameter, InterfaceTypeError)
-from interfaces import DateParameter, DateTimeParameter
 from noc.lib.validators import is_int
 from noc.models import is_document
 from noc.main.models.tag import Tag
+from noc.core.stencil import stencil_registry
+from .extapplication import ExtApplication, view
+from .interfaces import DateParameter, DateTimeParameter
 
 
 class ExtModelApplication(ExtApplication):
@@ -39,9 +41,10 @@ class ExtModelApplication(ExtApplication):
     query_condition = "startswith"  # Match method for string fields
     int_query_fields = []  # Query integer fields for exact match
     pk_field_name = None  # Set by constructor
-    clean_fields = {}  # field name -> Parameter instance
+    clean_fields = {"id": IntParameter()}  # field name -> Parameter instance
     custom_fields = {}  # name -> handler, populated automatically
     order_map = {}  # field name -> SQL query for ordering
+    lookup_default = [{"id": "Leave unchanged", "label": "Leave unchanged"}]
     ignored_fields = set(["id", "bi_id"])
 
     def __init__(self, *args, **kwargs):
@@ -255,6 +258,11 @@ class ExtModelApplication(ExtApplication):
             if f.name == "tags":
                 # Send tags as a list
                 r[f.name] = getattr(o, f.name)
+            elif f.name == "shape":
+                if o.shape:
+                    v = stencil_registry.get(o.shape)
+                    r[f.name] = v.id
+                    r["%s__label" % f.name] = unicode(v.title)
             elif hasattr(f, "document"):
                 # DocumentReferenceField
                 v = getattr(o, f.name)
@@ -266,8 +274,8 @@ class ExtModelApplication(ExtApplication):
                     r["%s__label" % f.name] = ""
             elif f.rel is None:
                 v = f._get_val_from_obj(o)
-                if (v is not None and type(v) not in (str, unicode, int, long, bool, list)):
-                    if type(v) == datetime.datetime:
+                if v is not None and not isinstance(v, (str, unicode, int, long, bool, list)):
+                    if isinstance(v, datetime.datetime):
                         v = v.isoformat()
                     else:
                         v = unicode(v)
@@ -355,9 +363,9 @@ class ExtModelApplication(ExtApplication):
                 fname = o[1:]
             else:
                 fname = o
-            if fname in self.order_map:
+            if o in self.order_map:
                 no = "%s_order_%d" % (fname, n)
-                extra_select[no] = self.order_map[fname]
+                extra_select[no] = self.order_map[o]
                 new_order += [no]
             else:
                 new_order += [o]
@@ -366,13 +374,46 @@ class ExtModelApplication(ExtApplication):
             extra["select"] = extra_select
         return extra, new_order
 
+    def can_create(self, user, obj):
+        """
+        Check user can create object. Used to additional
+        restrictions after permissions check
+        :param user:
+        :param obj: Object instance
+        :return: True if access granted
+        """
+        return True
+
+    def can_update(self, user, obj):
+        """
+        Check user can update object. Used to additional
+        restrictions after permissions check
+        :param user:
+        :param obj: Object instance
+        :return: True if access granted
+        """
+        return True
+
+    def can_delete(self, user, obj):
+        """
+        Check user can delete object. Used to additional
+        restrictions after permissions check
+        :param user:
+        :param obj: Object instance
+        :return: True if access granted
+        """
+        return True
+
     @view(method=["GET"], url="^$", access="read", api=True)
     def api_list(self, request):
         return self.list_data(request, self.instance_to_dict)
 
     @view(method=["GET"], url=r"^lookup/$", access="lookup", api=True)
     def api_lookup(self, request):
-        return self.list_data(request, self.instance_to_lookup)
+        try:
+            return self.list_data(request, self.instance_to_lookup)
+        except ValueError:
+            return self.response(self.lookup_default, status=self.OK)
 
     @view(method=["POST"], url="^$", access="create", api=True)
     def api_create(self, request):
@@ -410,7 +451,7 @@ class ExtModelApplication(ExtApplication):
             o = self.model(**attrs)
             # Run models validators
             try:
-                o.clean_fields()
+                o.full_clean(exclude=list(self.ignored_fields))
             except ValidationError as e:
                 e_msg = []
                 for f in e.message_dict:
@@ -419,6 +460,12 @@ class ExtModelApplication(ExtApplication):
                     "status": False,
                     "message": "Validation error: %s" % " | ".join(e_msg)
                 }, status=self.BAD_REQUEST)
+            # Check permissions
+            if not self.can_create(request.user, o):
+                return self.render_json({
+                    "status": False,
+                    "message": "Permission denied"
+                }, status=self.FORBIDDEN)
             # Check for duplicates
             try:
                 o.save()
@@ -482,10 +529,10 @@ class ExtModelApplication(ExtApplication):
             return HttpResponse("", status=self.NOT_FOUND)
         # Tags
         if hasattr(o, "tags") and attrs.get("tags"):
-            for t in set(getattr(o, "tags", [])) - (set(attrs.get("tags", []))):
+            for t in set(getattr(o, "tags") or []) - (set(attrs.get("tags", []))):
                 Tag.unregister_tag(t, repr(self.model))
                 self.logger.info("Unregister Tag: %s" % t)
-            for t in set(attrs.get("tags", [])) - (set(getattr(o, "tags", []))):
+            for t in set(attrs.get("tags", [])) - (set(getattr(o, "tags") or [])):
                 Tag.register_tag(t, repr(self.model))
                 self.logger.info("Register Tag: %s" % t)
         # Update attributes
@@ -502,6 +549,12 @@ class ExtModelApplication(ExtApplication):
                 "status": False,
                 "message": "Validation error: %s" % " | ".join(e_msg)
             }, status=self.BAD_REQUEST)
+        # Check permissions
+        if not self.can_create(request.user, o):
+            return self.render_json({
+                "status": False,
+                "message": "Permission denied"
+            }, status=self.FORBIDDEN)
         # Save
         try:
             o.save()
@@ -531,6 +584,12 @@ class ExtModelApplication(ExtApplication):
                 "status": False,
                 "message": "Not found"
             }, status=self.NOT_FOUND)
+        # Check permissions
+        if not self.can_delete(request.user, o):
+            return self.render_json({
+                "status": False,
+                "message": "Permission denied"
+            }, status=self.FORBIDDEN)
         try:
             o.delete()
         except ValueError as e:
